+++ backend/src/common/types/cars/autoria-request-params.ts
@@ -1,0 +1,12 @@
+interface AutoriaRequestParams {
+  category_id: number;
+  price_ot: string;
+  price_do: string;
+  powerFrom: string;
+  powerTo: string;
+  power_name: number;
+  engineVolumeFrom: string;
+  engineVolumeTo: string;
+}
+
+export type { AutoriaRequestParams };
