--- backend/src/common/types/cars/autoria_request_params.ts
@@ -1,12 +1,0 @@
-interface AutoriaRequestParams {
-  category_id: number;
-  price_ot?: string;
-  price_do?: string;
-  powerFrom?: string;
-  powerTo?: string;
-  power_name?: number;
-  engineVolumeFrom?: string;
-  engineVolumeTo?: string;
-}
-
-export type { AutoriaRequestParams };
